Move Set into its own package
