+++ set/set_test.go
@@ -1,6 +1,10 @@
-package sliceutils
+package set
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/Shresht7/sliceutils"
+)
 
 func TestSet(t *testing.T) {
 
@@ -8,7 +12,7 @@ func TestSet(t *testing.T) {
 
 	set := GetSet(slice)
 
-	if !Equal(set, []int{0, 1, 2, 3, 4, 5}) {
+	if !sliceutils.Equal(set, []int{0, 1, 2, 3, 4, 5}) {
 		t.Error("Failed to create set")
 	}
